Move RustupToolchainCheck to module within the check module
